requirements.txt
aiofiles
matplotlib
numpy
openai
opencv-python
pandas
pillow
python-dotenv
scikit-learn
tiktoken
tqdm
